# cmd/server/docs/swagger.yaml
definitions:
  schemas.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  schemas.CategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  schemas.ItemsCart:
    properties:
      name:
        example: Product 1
        type: string
      product_id:
        example: 1
        type: integer
      quantity:
        example: 1
        type: integer
      unit_price:
        example: 100
        type: number
    required:
    - name
    - product_id
    - quantity
    - unit_price
    type: object
  schemas.Payer:
    properties:
      email:
        example: johndoe@example.com
        type: string
      name:
        example: John
        type: string
      surname:
        example: Doe
        type: string
    required:
    - email
    - name
    - surname
    type: object
  schemas.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/schemas.CategoryResponse'
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      primary_image:
        type: string
      secondary_images:
        items:
          type: string
        type: array
      stock:
        type: number
    type: object
  schemas.ProductResponseDTO:
    properties:
      category:
        $ref: '#/definitions/schemas.CategoryResponse'
      code:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      primary_image:
        type: string
      stock:
        type: number
    type: object
  schemas.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  schemas.ShoppingCart:
    properties:
      client:
        $ref: '#/definitions/schemas.Payer'
      items:
        items:
          $ref: '#/definitions/schemas.ItemsCart'
        minItems: 1
        type: array
    required:
    - client
    - items
    type: object
  schemas.TenantResponse:
    properties:
      expiration:
        type: string
      identifier:
        type: string
      is_active:
        type: boolean
    type: object
info:
  contact:
    email: danielmchachagua@gmail.com
    name: Daniel Chachagua
  description: This is a api to app noa gestion ecommerce microservice.
  termsOfService: http://swagger.io/terms/
  title: APP NOA Gestion Ecommerce API
  version: "1.0"
paths:
  /ecommerce/{tenantID}/api/v1/category/get_all:
    get:
      consumes:
      - application/json
      description: Obtener todas las categorías disponibles
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.Category'
                  type: array
              type: object
      summary: CategoryGetAll
      tags:
      - Category
  /ecommerce/{tenantID}/api/v1/image/get/{filename}:
    get:
      consumes:
      - application/json
      description: Obtener imagen
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      - description: codigo del producto
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: ImageGet
      tags:
      - Image
  /ecommerce/{tenantID}/api/v1/mp/generate_link:
    post:
      consumes:
      - application/json
      description: Obtener link de pago
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      - description: Body
        in: body
        name: shopping_cart
        required: true
        schema:
          $ref: '#/definitions/schemas.ShoppingCart'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
      summary: MPGenerateLink
      tags:
      - Mercado Pago
  /ecommerce/{tenantID}/api/v1/mp/info_pay/{reference}:
    get:
      consumes:
      - application/json
      description: Obtener información de pago
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      - description: Referencia de pago
        in: path
        name: reference
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
      summary: MPInfoPay
      tags:
      - Mercado Pago
  /ecommerce/{tenantID}/api/v1/mp/state_pay:
    post:
      consumes:
      - application/json
      description: Obtener link de pago
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
      summary: MPStatePay
      tags:
      - Mercado Pago
  /ecommerce/{tenantID}/api/v1/product/get_by_code:
    get:
      consumes:
      - application/json
      description: Obtener un producto por su código
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      - description: codigo del producto
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.ProductResponse'
              type: object
      summary: ProductGetByCode
      tags:
      - Product
  /ecommerce/{tenantID}/api/v1/product/get_page:
    get:
      consumes:
      - application/json
      description: Obtener un producto por su código
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      - description: pagina
        in: query
        name: page
        type: integer
      - description: limite
        in: query
        name: limit
        type: integer
      - description: nombre del producto
        in: query
        name: name
        type: string
      - description: categoria id del producto
        in: query
        name: category_id
        type: integer
      - description: ordenar por
        enum:
        - PRICE_LOW_TO_HIGH
        - PRICE_HIGH_TO_LOW
        - NAME_A_Z
        - NAME_Z_A
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.ProductResponseDTO'
              type: object
      summary: ProductGetByCode
      tags:
      - Product
  /ecommerce/{tenantID}/api/v1/product/upload_image:
    post:
      consumes:
      - multipart/form-data
      description: |-
        ### Flujo de Carga de Imágenes
        Genera un token temporal desde la API principal para subir imágenes al microservicio.

        **Endpoitn de la api princial para pedir token:**
        ~~~
        POST /api/v1/product/generate_token_to_image
        ~~~

        **Pasos requeridos:**
        1. Incluir el token en el header `x-token-tenant`.

        > *Nota: El token tiene una validez limitada de 30 minutos.*
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      - description: Token o ID de validación personalizada
        in: header
        name: x-token-tenant
        required: true
        type: string
      - description: Imagen principal del producto
        in: formData
        name: primaryImage
        type: file
      - collectionFormat: csv
        description: Imágenes secundarias (puedes enviar varias)
        in: formData
        items:
          type: file
        name: secondaryImage
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
      summary: ProductSaveImage
      tags:
      - Product
  /ecommerce/{tenantID}/api/v1/tenant/get:
    get:
      consumes:
      - application/json
      description: Obtener tenant
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.TenantResponse'
                  type: array
              type: object
      summary: TenantGet
      tags:
      - Tenant
  /ecommerce/{tenantID}/api/v1/tenant/upload_image:
    post:
      consumes:
      - multipart/form-data
      description: |-
        ### Flujo de Carga de Imágenes
        Genera un token temporal desde la API principal para subir imágenes al microservicio.

        **Endpoitn de la api princial para pedir token:**
        ~~~
        POST /api/v1/tenant/generate_token_to_image_setting
        ~~~

        **Pasos requeridos:**
        1. Incluir el token en el header `x-token-tenant`.

        > *Nota: El token tiene una validez limitada de 30 minutos.*
      parameters:
      - description: ID del Tenant
        in: path
        name: tenantID
        required: true
        type: string
      - description: Token o ID de validación personalizada
        in: header
        name: x-token-tenant
        required: true
        type: string
      - description: Imagen del logo
        in: formData
        name: logoImage
        type: file
      - description: Imágen de la portada
        in: formData
        name: frontPageImage
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
      summary: TenantSaveImage
      tags:
      - Tenant
swagger: "2.0"